server/requirements.txt
aiofiles==0.4.0
asyncpg==0.18.3
databases==0.2.2
fastapi==0.23.0
itsdangerous==1.1.0
Jinja2==2.10.1
Pillow==6.0.0
pydantic==0.25
SQLAlchemy==1.3.3
aioredis==1.2.0